requirements.txt
blinker==1.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
pandas==1.0.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
xlrd==1.2.0
